model/auth.goをapi/auth.goへ移動
