+++ server/internal/model/user.go
@@ -1,0 +1,14 @@
+package model
+
+import (
+
+
+	// "github.com/jmoiron/sqlx"
+)
+
+type User struct {
+	UserId int64 `db:"user_id"`
+	UserName string `db:"user_name"`
+}
+
+
