+++ server/internal/routers/api/auth.go
@@ -1,0 +1,13 @@
+package api
+
+type Auth struct {
+	ID       string `json:"id"`
+	Password string `json:"password"`
+}
+
+// ユーザー登録
+func Account(userId, password string) (bool, error) {
+	var auth Auth
+	err := db.Select("id")
+	// 既に同じ名前が使われている場合
+}
